fix(desktop): surface model requests the harness rejected
When a requested model has no byte-exact match in the harness catalog,
buzz-acp warns and starts the session on the agent's own default instead.
The only signal reaching desktop was `modelOverridden`, which is false
both when a request was rejected AND when no model was requested at all,
so the ModelPicker kept displaying the rejected pick as though it had
taken effect. A wrong model was indistinguishable from a right one, with
nothing surfaced anywhere.

The harness now reports `requestedModel` and `modelApplied` alongside
`modelOverridden` in `session_config_captured`. The config surface derives
`unappliedModelRequest` from that pair, and the ModelPicker renders
"<model> not applied" when the live session is not honouring the pick.

Both fields are purely additive; existing `model_overridden` consumers are
untouched. Harnesses that predate the pair omit both, and a missing
`requestedModel` keeps those sessions out of the warning rather than
mislabelling them.

Two mechanical extractions are included because
`desktop/scripts/check-file-sizes.mjs` forbids *any* growth in files
already past the 1000-line cap, and both files this fix must touch were
over it:

  - `commands/agent_config.rs` (1112) — ACP session-payload parsing moved
    verbatim to `config_bridge/session_payload.rs`, beside the types it
    produces. Now 989.
  - `shared/api/types.ts` (1032) — config-bridge types moved verbatim to
    `shared/api/configBridge.ts` and re-exported. Now 956.

Both moves are code-identical with no behaviour change.

Refs #2692, #4004, #2265
